backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
pandas
openpyxl
requests
jinja2
python-multipart